requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
annotated-doc
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
async-timeout
attrs
Authlib
banks
bcrypt
beautifulsoup4
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
dataclasses-json
Deprecated
deprecation
dirtyjson
dnspython
ecdsa
email-validator
exceptiongroup
fastapi
filelock
filetype
frozenlist
fsspec
google-ai-generativelanguage
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-genai
googleapis-common-protos
greenlet
griffe
grpcio
grpcio-status
h11
hf-xet
httpcore
httplib2
httpx
httpx-sse
huggingface-hub
idna
Jinja2
joblib
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith
llama-cloud
llama-cloud-services
llama-index-core
llama-index-instrumentation
llama-index-workflows
llama-parse
MarkupSafe
marshmallow
mpmath
multidict
mypy_extensions
nest-asyncio
networkx
nltk
numpy
oauthlib
orjson
openpyxl
packaging
pandas
passlib
pdfminer.six
pdfplumber
pillow
platformdirs
propcache
proto-plus
protobuf
psycopg2
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
PyMuPDF
pyparsing
pypdfium2
pytesseract
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
py7zr
rarfile
regex
requests
requests-oauthlib
requests-toolbelt
rsa
safetensors
scikit-learn
scipy
sentence-transformers
six
SQLAlchemy
sse-starlette
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
typing-inspect
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
validators
weaviate
weaviate-client
wrapt
yarl
zstandard